backend/requirements.txt
fastapi==0.111.1
langchain==0.2.10
langchain_chroma==0.1.2
langchain_community==0.2.9
langchain_core==0.2.22
langchain_google_vertexai==1.0.6
langchain_text_splitters==0.2.2
protobuf==3.12.4
pydantic==2.8.2
pymongo==4.6.1
python-dotenv==1.0.1
web3==6.20.0
google-cloud-speech